+++ tests/__init__.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+
+testcase = TestCase()
+ae = testcase.assertEqual